requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic-settings==2.6.0
aiosqlite==0.20.0
PyMuPDF==1.25.3
openpyxl==3.1.5
google-genai==1.32.0
